Renamed models to match node style
