+++ src/client/app/services/player.service.ts
@@ -1,7 +1,7 @@
 
 import { Injectable } from '@angular/core';
 import { RestService } from "./rest.service";
-import { Player } from "shared/models/player";
+import { Player } from "shared/models";
 
 @Injectable()
 export class PlayerService {
