+++ src/server/app/models/GameModel.ts
@@ -1,6 +1,6 @@
 
 import * as mongoose from "mongoose";
-import { IGame } from "shared/models/IGame";
+import { IGame } from "shared/models";
 
 export interface GameModel extends IGame, mongoose.Document {
