+++ src/server/app/models/PlayerModel.ts
@@ -1,6 +1,6 @@
 
 import * as mongoose from "mongoose";
-import { IPlayer } from "shared/models/IPlayer";
+import { IPlayer } from "shared/models";
 
 export interface PlayerModel extends IPlayer, mongoose.Document {
