+++ src/shared/models.ts
@@ -1,0 +1,2 @@
+
+export * from "./models/index";
