+++ src/shared/models/Faction.ts
@@ -1,0 +1,6 @@
+
+
+export enum Faction {
+    Knights,
+    Minions
+}
