+++ src/shared/models/GameOptions.ts
