+++ src/shared/models/GamePlayData.ts
@@ -1,5 +1,5 @@
 
-import { Role } from "./role";
+import { Role } from "./Role";
 
 export class GamePlayData {
     gameId: string;
