+++ src/shared/models/IGame.ts
@@ -1,7 +1,7 @@
 
-import { Player } from "./player";
-import { GameOptions } from "./game-options";
-import { GamePlayData } from "./game-play-data";
+import { Player } from "./Player";
+import { GameOptions } from "./GameOptions";
+import { GamePlayData } from "./GamePlayData";
 
 export interface IGame {
     name: string;
