+++ src/shared/models/PlayerInfo.ts
