+++ src/shared/models/index.ts
@@ -1,0 +1,13 @@
+
+
+export { Faction } from "./Faction";
+export { Game } from "./Game";
+export { GameLobbyData } from "./GameLobbyData";
+export { GameOptions } from "./GameOptions";
+export { GamePlayData } from "./GamePlayData";
+export { IGame } from "./IGame";
+export { IIndexable } from "./IIndexable";
+export { IPlayer } from "./IPlayer";
+export { Player } from "./Player";
+export { PlayerInfo } from "./PlayerInfo";
+export { Role } from "./Role";
